extract components from index page
