+++ frontend/components/Disclaimer.vue
@@ -1,0 +1,20 @@
+<template>
+  <section
+    class="
+      bg-gray-200
+      pl-4
+      py-5
+      md:w-4/5 md:mx-auto
+      border-indigo-600 border-l-8
+      rounded-md
+      text-indigo-600
+    "
+  >
+    <h2 class="font-bold text-lg">Disclaimer</h2>
+    <p class="my-3 italic">
+      This is still a prototype under constant evolution. There is a list of
+      pre-condition we require to guarantee proper work.
+    </p>
+    <nuxt-link class="underline" to="/">learn more ></nuxt-link>
+  </section>
+</template>
