+++ frontend/components/ProtoHeader.vue
@@ -1,0 +1,5 @@
+<template>
+  <header>
+    <h1 class="text-4xl font-bold text-center my-10">ProtoINCT</h1>
+  </header>
+</template>
